+++ lib/Elastica/Query/ConstantScore.php
@@ -2,7 +2,6 @@
 namespace Elastica\Query;
 
 use Elastica\Exception\InvalidException;
-use Elastica\Filter\AbstractFilter;
 
 /**
  * Constant score query.
@@ -18,15 +17,9 @@ class ConstantScore extends AbstractQuery
      *
      * @param null|\Elastica\Query\AbstractQuery|array $filter
      */
-    public function __construct($filter = null)
+    public function __construct(AbstractQuery $filter = null)
     {
         if (!is_null($filter)) {
-            if ($filter instanceof AbstractFilter) {
-                trigger_error('Deprecated: Elastica\Query\ConstantScore passing AbstractFilter is deprecated. Pass AbstractQuery instead.', E_USER_DEPRECATED);
-            } elseif (!is_array($filter) && !($filter instanceof AbstractQuery)) {
-                throw new InvalidException('Filter must be instance of AbstractQuery');
-            }
-
             $this->setFilter($filter);
         }
     }
@@ -38,14 +31,8 @@ class ConstantScore extends AbstractQuery
      *
      * @return $this
      */
-    public function setFilter($filter)
+    public function setFilter(AbstractQuery $filter)
     {
-        if ($filter instanceof AbstractFilter) {
-            trigger_error('Deprecated: Elastica\Query\ConstantScore::setFilter passing AbstractFilter is deprecated. Pass AbstractQuery instead.', E_USER_DEPRECATED);
-        } elseif (!is_array($filter) && !($filter instanceof AbstractQuery)) {
-            throw new InvalidException('Filter must be instance of AbstractQuery or array');
-        }
-
         return $this->setParam('filter', $filter);
     }