+++ lib/Elastica/Query/Filtered.php
@@ -2,7 +2,6 @@
 namespace Elastica\Query;
 
 use Elastica\Exception\InvalidException;
-use Elastica\Filter\AbstractFilter;
 
 trigger_error('Use BoolQuery instead. Filtered query is deprecated since ES 2.0.0-beta1 and this class will be removed in further Elastica releases.', E_USER_DEPRECATED);
 
@@ -23,18 +22,9 @@ class Filtered extends AbstractQuery
      * @param \Elastica\Query\AbstractQuery $query  OPTIONAL Query object
      * @param \Elastica\Query\AbstractQuery $filter OPTIONAL Filter object
      */
-    public function __construct(AbstractQuery $query = null, $filter = null)
+    public function __construct(AbstractQuery $query = null, AbstractQuery $filter = null)
     {
         $this->setQuery($query);
-
-        if (null !== $filter) {
-            if ($filter instanceof AbstractFilter) {
-                trigger_error('Deprecated: Elastica\Query\Filtered passing AbstractFilter is deprecated. Pass AbstractQuery instead.', E_USER_DEPRECATED);
-            } elseif (!($filter instanceof AbstractQuery)) {
-                throw new InvalidException('Filter must be instance of AbstractQuery');
-            }
-        }
-
         $this->setFilter($filter);
     }
 
@@ -57,23 +47,15 @@ class Filtered extends AbstractQuery
      *
      * @return $this
      */
-    public function setFilter($filter = null)
+    public function setFilter(AbstractQuery $filter = null)
     {
-        if (null !== $filter) {
-            if ($filter instanceof AbstractFilter) {
-                trigger_error('Deprecated: Elastica\Query\Filtered::setFilter passing AbstractFilter is deprecated. Pass AbstractQuery instead.', E_USER_DEPRECATED);
-            } elseif (!($filter instanceof AbstractQuery)) {
-                throw new InvalidException('Filter must be instance of AbstractQuery');
-            }
-        }
-
         return $this->setParam('filter', $filter);
     }
 
     /**
      * Gets the filter.
      *
-     * @return \Elastica\Query\AbstractQuery|\Elastica\Filter\AbstractFilter
+     * @return \Elastica\Query\AbstractQuery
      */
     public function getFilter()
     {
@@ -105,7 +87,7 @@ class Filtered extends AbstractQuery
             $filtered['query'] = $this->getParam('query')->toArray();
         }
 
-        if ($this->hasParam('filter') && ($this->getParam('filter') instanceof AbstractQuery || $this->getParam('filter') instanceof AbstractFilter)) {
+        if ($this->hasParam('filter') && $this->getParam('filter') instanceof AbstractQuery) {
             $filtered['filter'] = $this->getParam('filter')->toArray();
         }