+++ lib/Elastica/Search.php
@@ -2,7 +2,6 @@
 namespace Elastica;
 
 use Elastica\Exception\InvalidException;
-use Elastica\Filter\AbstractFilter;
 use Elastica\ResultSet\BuilderInterface;
 use Elastica\ResultSet\DefaultBuilder;
 
@@ -176,10 +175,6 @@ class Search
      */
     public function setQuery($query)
     {
-        if ($query instanceof AbstractFilter) {
-            trigger_error('Deprecated: Elastica\Search::setQuery() passing AbstractFilter is deprecated. Create query and use setPostFilter with AbstractQuery instead.', E_USER_DEPRECATED);
-        }
-
         $this->_query = Query::create($query);
 
         return $this;