+++ test/lib/Elastica/Test/Query/FunctionScoreTest.php
@@ -2,10 +2,9 @@
 namespace Elastica\Test\Query;
 
 use Elastica\Document;
-use Elastica\Filter\Exists;
-use Elastica\Filter\Term;
 use Elastica\Query\FunctionScore;
 use Elastica\Query\MatchAll;
+use Elastica\Query\Term;
 use Elastica\Script\Script;
 use Elastica\Test\Base as BaseTest;
 
@@ -43,253 +42,6 @@ class FunctionScoreTest extends BaseTest
         $index->refresh();
 
         return $index;
-    }
-
-    /**
-     * @group unit
-     * @expectedException \Elastica\Exception\DeprecatedException
-     */
-    public function testSetFilterWithLegacyFilterDeprecated()
-    {
-        $this->hideDeprecated();
-        $existsFilter = new Exists('test');
-        $this->showDeprecated();
-
-        $query = new FunctionScore('test');
-
-        $query->setFilter($existsFilter);
-    }
-
-    /**
-     * @group unit
-     * @expectedException \Elastica\Exception\InvalidException
-     */
-    public function testAddFunctionInvalid()
-    {
-        $query = new FunctionScore('test');
-        $query->addFunction('f', 1, $this);
-    }
-
-    /**
-     * @group unit
-     */
-    public function testAddFunctionWithLegacyFilterDeprecated()
-    {
-        $this->hideDeprecated();
-        $existsFilter = new Exists('test');
-        $this->showDeprecated();
-
-        $query = new FunctionScore('test');
-
-        $errorsCollector = $this->startCollectErrors();
-        $query->addFunction('f', 1, $existsFilter);
-        $this->finishCollectErrors();
-
-        $errorsCollector->assertOnlyDeprecatedErrors(
-            [
-                'Deprecated: Elastica\Query\FunctionScore::addFunction passing AbstractFilter is deprecated. Pass AbstractQuery instead.',
-            ]
-        );
-    }
-
-    /**
-     * @group unit
-     * @expectedException \Elastica\Exception\InvalidException
-     */
-    public function testAddDecayFunctionInvalid()
-    {
-        $query = new FunctionScore('test');
-        $query->addDecayFunction(FunctionScore::DECAY_GAUSS, 'location', $this->locationOrigin, '2mi', null, null, null, $this);
-    }
-
-    /**
-     * @group unit
-     */
-    public function testAddDecayFunctionWithLegacyFilterDeprecated()
-    {
-        $this->hideDeprecated();
-        $existsFilter = new Exists('test');
-        $this->showDeprecated();
-
-        $query = new FunctionScore('test');
-
-        $errorsCollector = $this->startCollectErrors();
-        $query->addDecayFunction(FunctionScore::DECAY_GAUSS, 'location', $this->locationOrigin, '2mi', null, null, null, $existsFilter);
-        $this->finishCollectErrors();
-
-        $errorsCollector->assertOnlyDeprecatedErrors(
-            [
-                'Deprecated: Elastica\Query\FunctionScore::addDecayFunction passing AbstractFilter is deprecated. Pass AbstractQuery instead.',
-                'Deprecated: Elastica\Query\FunctionScore::addFunction passing AbstractFilter is deprecated. Pass AbstractQuery instead.',
-            ]
-        );
-    }
-
-    /**
-     * @group unit
-     * @expectedException \Elastica\Exception\InvalidException
-     */
-    public function testScriptScoreFunctionInvalid()
-    {
-        $query = new FunctionScore('test');
-        $query->addScriptScoreFunction(new Script('t'), $this);
-    }
-
-    /**
-     * @group unit
-     */
-    public function testScriptScoreFunctionWithLegacyFilterDeprecated()
-    {
-        $this->hideDeprecated();
-        $existsFilter = new Exists('test');
-        $this->showDeprecated();
-
-        $query = new FunctionScore('test');
-
-        $errorsCollector = $this->startCollectErrors();
-        $query->addScriptScoreFunction(new Script('t'), $existsFilter);
-        $this->finishCollectErrors();
-
-        $errorsCollector->assertOnlyDeprecatedErrors(
-            [
-                'Deprecated: Elastica\Query\FunctionScore::addScriptScoreFunction passing AbstractFilter is deprecated. Pass AbstractQuery instead.',
-                'Deprecated: Elastica\Query\FunctionScore::addFunction passing AbstractFilter is deprecated. Pass AbstractQuery instead.',
-            ]
-        );
-    }
-
-    /**
-     * @group unit
-     * @expectedException \Elastica\Exception\InvalidException
-     */
-    public function testAddFieldValueFactorFunctionInvalid()
-    {
-        $query = new FunctionScore('test');
-        $query->addFieldValueFactorFunction('popularity', 1.2, FunctionScore::FIELD_VALUE_FACTOR_MODIFIER_SQRT, 0.1, null, $this);
-    }
-
-    /**
-     * @group unit
-     */
-    public function testAddFieldValueFactorFunctionWithLegacyFilterDeprecated()
-    {
-        $this->hideDeprecated();
-        $existsFilter = new Exists('test');
-        $this->showDeprecated();
-
-        $query = new FunctionScore('test');
-
-        $errorsCollector = $this->startCollectErrors();
-        $query->addFieldValueFactorFunction('popularity', 1.2, FunctionScore::FIELD_VALUE_FACTOR_MODIFIER_SQRT, 0.1, null, $existsFilter);
-        $this->finishCollectErrors();
-
-        $errorsCollector->assertOnlyDeprecatedErrors(
-            [
-                'Deprecated: Elastica\Query\FunctionScore::addFieldValueFactorFunction passing AbstractFilter is deprecated. Pass AbstractQuery instead.',
-                'Deprecated: Elastica\Query\FunctionScore::addFunction passing AbstractFilter is deprecated. Pass AbstractQuery instead.',
-            ]
-        );
-    }
-
-    /**
-     * @group unit
-     * @expectedException \Elastica\Exception\InvalidException
-     */
-    public function testAddBoostFactorFunctionFunctionInvalid()
-    {
-        $query = new FunctionScore('test');
-        $query->addBoostFactorFunction(5.0, $this);
-    }
-
-    /**
-     * @group unit
-     */
-    public function testAddBoostFactorFunctionWithLegacyFilterDeprecated()
-    {
-        $this->hideDeprecated();
-        $existsFilter = new Exists('test');
-        $this->showDeprecated();
-
-        $query = new FunctionScore('test');
-
-        $errorsCollector = $this->startCollectErrors();
-        $query->addBoostFactorFunction(5.0, $existsFilter);
-        $this->finishCollectErrors();
-
-        $errorsCollector->assertOnlyDeprecatedErrors(
-            [
-                'Deprecated: Elastica\Query\FunctionScore::addBoostFactorFunction passing AbstractFilter is deprecated. Pass AbstractQuery instead.',
-                'Query\FunctionScore::addBoostFactorFunction is deprecated. Use addWeightFunction instead. This method will be removed in further Elastica releases',
-                'Deprecated: Elastica\Query\FunctionScore::addWeightFunction passing AbstractFilter is deprecated. Pass AbstractQuery instead.',
-                'Deprecated: Elastica\Query\FunctionScore::addFunction passing AbstractFilter is deprecated. Pass AbstractQuery instead.',
-            ]
-        );
-    }
-
-    /**
-     * @group unit
-     * @expectedException \Elastica\Exception\InvalidException
-     */
-    public function testAddWeightFunctionFunctionInvalid()
-    {
-        $query = new FunctionScore('test');
-        $query->addWeightFunction(5.0, $this);
-    }
-
-    /**
-     * @group unit
-     */
-    public function testAddWeightFunctionWithLegacyFilterDeprecated()
-    {
-        $this->hideDeprecated();
-        $existsFilter = new Exists('test');
-        $this->showDeprecated();
-
-        $query = new FunctionScore('test');
-
-        $errorsCollector = $this->startCollectErrors();
-        $query->addWeightFunction(5.0, $existsFilter);
-        $this->finishCollectErrors();
-
-        $errorsCollector->assertOnlyDeprecatedErrors(
-            [
-                'Deprecated: Elastica\Query\FunctionScore::addWeightFunction passing AbstractFilter is deprecated. Pass AbstractQuery instead.',
-                'Deprecated: Elastica\Query\FunctionScore::addFunction passing AbstractFilter is deprecated. Pass AbstractQuery instead.',
-            ]
-        );
-    }
-
-    /**
-     * @group unit
-     * @expectedException \Elastica\Exception\InvalidException
-     */
-    public function testAddRandomScoreFunctionInvalid()
-    {
-        $query = new FunctionScore('test');
-        $query->addRandomScoreFunction(5.0, $this);
-    }
-
-    /**
-     * @group unit
-     */
-    public function testAddRandomScoreFunctionWithLegacyFilterDeprecated()
-    {
-        $this->hideDeprecated();
-        $existsFilter = new Exists('test');
-        $this->showDeprecated();
-
-        $query = new FunctionScore('test');
-
-        $errorsCollector = $this->startCollectErrors();
-        $query->addRandomScoreFunction(5.0, $existsFilter);
-        $this->finishCollectErrors();
-
-        $errorsCollector->assertOnlyDeprecatedErrors(
-            [
-                'Deprecated: Elastica\Query\FunctionScore::addRandomScoreFunction passing AbstractFilter is deprecated. Pass AbstractQuery instead.',
-                'Deprecated: Elastica\Query\FunctionScore::addFunction passing AbstractFilter is deprecated. Pass AbstractQuery instead.',
-            ]
-        );
     }
 
     /**
@@ -438,7 +190,7 @@ class FunctionScoreTest extends BaseTest
     /**
      * @group unit
      */
-    public function testAddBoostFactorFunction()
+    public function testAddWeightFunction()
     {
         $filter = new \Elastica\Query\Term(['price' => 4.5]);
         $query = new FunctionScore();
@@ -447,7 +199,7 @@ class FunctionScoreTest extends BaseTest
         $sameFilter = new \Elastica\Query\Term(['price' => 4.5]);
         $sameQuery = new FunctionScore();
         $this->hideDeprecated();
-        $sameQuery->addBoostFactorFunction(5.0, $sameFilter);
+        $sameQuery->addWeightFunction(5.0, $sameFilter);
         $this->showDeprecated();
 
         $this->assertEquals($query->toArray(), $sameQuery->toArray());
@@ -456,7 +208,7 @@ class FunctionScoreTest extends BaseTest
     /**
      * @group unit
      */
-    public function testLegacyFilterAddBoostFactorFunction()
+    public function testLegacyFilterAddWeightFunction()
     {
         $query = new FunctionScore();
         $this->hideDeprecated();
@@ -467,7 +219,7 @@ class FunctionScoreTest extends BaseTest
         $sameQuery = new FunctionScore();
         $this->hideDeprecated();
         $sameFilter = new Term(['price' => 4.5]);
-        $sameQuery->addBoostFactorFunction(5.0, $sameFilter);
+        $sameQuery->addWeightFunction(5.0, $sameFilter);
         $this->showDeprecated();
 
         $this->assertEquals($query->toArray(), $sameQuery->toArray());