+++ test/lib/Elastica/Test/QueryBuilder/DSL/QueryTest.php
@@ -1,9 +1,7 @@
 <?php
 namespace Elastica\Test\QueryBuilder\DSL;
 
-use Elastica\Filter\Exists;
 use Elastica\Query\Match;
-use Elastica\Query\Term;
 use Elastica\QueryBuilder\DSL;
 
 class QueryTest extends AbstractDSLTest
@@ -33,59 +31,6 @@ class QueryTest extends AbstractDSLTest
 
     /**
      * @group unit
-     * @expectedException \Elastica\Exception\InvalidException
-     */
-    public function testConstantScoreFilterInvalid()
-    {
-        $queryDSL = new DSL\Query();
-        $queryDSL->constant_score($this);
-    }
-
-    /**
-     * @group unit
-     */
-    public function testConstantScoreWithLegacyFilterDeprecated()
-    {
-        $this->hideDeprecated();
-        $existsFilter = new Exists('test');
-        $this->showDeprecated();
-
-        $queryDSL = new DSL\Query();
-
-        $errorsCollector = $this->startCollectErrors();
-        $queryDSL->constant_score($existsFilter);
-        $this->finishCollectErrors();
-
-        $errorsCollector->assertOnlyDeprecatedErrors(
-            [
-                'Deprecated: Elastica\Query\ConstantScore passing AbstractFilter is deprecated. Pass AbstractQuery instead.',
-                'Deprecated: Elastica\Query\ConstantScore::setFilter passing AbstractFilter is deprecated. Pass AbstractQuery instead.',
-            ]
-        );
-    }
-
-    /**
-     * @group unit
-     */
-    public function testFilteredDeprecated()
-    {
-        $errorsCollector = $this->startCollectErrors();
-
-        $queryDSL = new DSL\Query();
-        $queryDSL->filtered(null, new Exists('term'));
-        $this->finishCollectErrors();
-
-        $errorsCollector->assertOnlyDeprecatedErrors(
-            [
-                'Use bool() instead. Filtered query is deprecated since ES 2.0.0-beta1 and this method will be removed in further Elastica releases.',
-                'Deprecated: Elastica\Query\Filtered passing AbstractFilter is deprecated. Pass AbstractQuery instead.',
-                'Deprecated: Elastica\Query\Filtered::setFilter passing AbstractFilter is deprecated. Pass AbstractQuery instead.',
-            ]
-        );
-    }
-
-    /**
-     * @group unit
      */
     public function testInterface()
     {
@@ -94,14 +39,7 @@ class QueryTest extends AbstractDSLTest
         $this->_assertImplemented($queryDSL, 'bool', 'Elastica\Query\BoolQuery', []);
         $this->_assertImplemented($queryDSL, 'boosting', 'Elastica\Query\Boosting', []);
         $this->_assertImplemented($queryDSL, 'common_terms', 'Elastica\Query\Common', ['field', 'query', 0.001]);
-        $this->_assertImplemented($queryDSL, 'constant_score', 'Elastica\Query\ConstantScore', [new Match()]);
         $this->_assertImplemented($queryDSL, 'dis_max', 'Elastica\Query\DisMax', []);
-
-        $this->hideDeprecated();
-        $this->_assertImplemented($queryDSL, 'filtered', 'Elastica\Query\Filtered', [new Match(), new Exists('field')]);
-        $this->_assertImplemented($queryDSL, 'filtered', 'Elastica\Query\Filtered', [new Match(), new Term()]);
-        $this->showDeprecated();
-
         $this->_assertImplemented($queryDSL, 'function_score', 'Elastica\Query\FunctionScore', []);
         $this->_assertImplemented($queryDSL, 'fuzzy', 'Elastica\Query\Fuzzy', ['field', 'type']);
         $this->_assertImplemented($queryDSL, 'has_child', 'Elastica\Query\HasChild', [new Match()]);