+++ test/lib/Elastica/Test/SuggestTest.php
@@ -2,7 +2,7 @@
 namespace Elastica\Test;
 
 use Elastica\Exception\NotImplementedException;
-use Elastica\Filter\BoolFilter;
+use Elastica\Query\BoolQuery;
 use Elastica\Suggest;
 use Elastica\Test\Base as BaseTest;
 
@@ -44,7 +44,7 @@ class SuggestTest extends BaseTest
     public function testCreateWithNonSuggest()
     {
         try {
-            Suggest::create(new BoolFilter());
+            Suggest::create(new BoolQuery());
             $this->fail();
         } catch (NotImplementedException $e) {
         }